Ventilate nav menus
As far as ventilating menus is concerned, our options are quite
limited by the fact that the WordPress importer will only import nav
menu items that go with posts (of any type) that are being ingested at
the same time. It is not possible, for instance, to synthesize a grand
unified menu in a separate WXR file and ingest that.

* Discard all menus except "main", "main-en", "main-fr" etc.

* Nest the contents of these under a new nav menu item, to make the
  whole submenu easier to manipulate in the wp-admin area. Rather than
  having to pick up the pieces from the top level, one can move the
  new submenu as a whole, and delete it once it is in the proper
  place (if desired)